requirements.txt
pillow==11.0.0
PyPDF2==3.0.1
